Implemented the last submenus.
This may have fixed issue #28, because it is now no longer necessary to scroll on the iPad, I believe. However, I cannot confirm this as I do not have access to the iPad right now.

While implementing the last submenus, I also styled everything a bit. In my opinion, the app is now more polished.